Fireball
Created a fireball that can be used by any entity. Fireball plays an animation and is angled towards some target position. Added a spell_cast state to the player so the fireball can be used by pressing the "F" key. The fireball will then move towards the current position of the mouse.